requirements.txt
flask==3.0.0
flask-cors==6.0.1
gunicorn==21.2.0
google-api-python-client==2.119.0
google-auth==2.27.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
requests==2.31.0
